# FastFB-Video-Downloader/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface FacebookVideoInput {
  /** @minLength 1 */
  url: string;
}

export interface FacebookDownloadInput {
  /** @minLength 1 */
  url: string;
  /** @nullable */
  formatId?: string | null;
}

export interface FacebookVideoFormat {
  formatId: string;
  label: string;
  /** @nullable */
  height: number | null;
  /** @nullable */
  filesize: number | null;
  /** @nullable */
  filesizeReadable: string | null;
  ext: string;
}

export interface FacebookVideoInfo {
  title: string;
  /** @nullable */
  thumbnail: string | null;
  /** @nullable */
  duration: number | null;
  /** @nullable */
  uploader: string | null;
  formats: FacebookVideoFormat[];
}